pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "allmight"
version = "0.1.0"
description = "Active Knowledge Graph Framework — the active layer on top of SMAK"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "tl32rodan" }]

dependencies = [
    "click>=8.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.scripts]
allmight = "allmight.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/allmight"]

[tool.pytest.ini_options]
testpaths = ["tests"]
